% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCOBBS.R
\docType{package}
\name{COBBS-package}
\alias{COBBS-package}
\alias{COBBS}
\title{Continuous Optimization Benchmarks By Simulation}
\description{
Continuous Optimization Benchmarks By Simulation
}
\details{
This package implements and demonstrates
the use of simulation with Gaussian process regression models
for benchmarking of continuous optimization algorithms.

\tabular{ll}{
Package: \tab COBBS\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2020-04-23\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\seealso{
Interface of main function: \code{\link{generateCOBBS}}
}
\author{
Martin Zaefferer \email{mzaefferer@gmail.com}
}
\keyword{package}
