% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotBenchmarkPerformance}
\alias{plotBenchmarkPerformance}
\title{Performance plot}
\usage{
plotBenchmarkPerformance(data, names = NULL)
}
\arguments{
\item{data}{a list of data frames created with \code{\link{loggedExperiment}}. Each data frame represents the result of an experiment with a different objective function.}

\item{names}{a character vector. Each represents a name for the underlying objective function of the \code{data} list.}
}
\value{
a ggplot instance
}
\description{
Create a simple performance plot, showing the scaled performance of an algorithm over time (based on experiments with different objective functions).
}
\examples{
## create data by running optim L-BFGS-B with two objective functions
lower <- -1
upper <- 1
require(nloptr)
expr <- expression(
  res <- lbfgs(x0=runif(1,lower,upper),fn = fnlog,
                     lower=lower,upper=upper)
)
res1 <- loggedExperiment(expr, function(x)x^2, 1:10)
res2 <- loggedExperiment(expr, function(x)sqrt(abs(x)), 1:10)
## create the plot from the results
p <- plotBenchmarkPerformance(list(res1,res2),names=c("square","root"))
print(p)

}
