% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianProcessRegression.R
\name{print.cobbsGPR}
\alias{print.cobbsGPR}
\title{Print Function GPR}
\usage{
\method{print}{cobbsGPR}(x, ...)
}
\arguments{
\item{x}{fit returned by \code{\link{gaussianProcessRegression}}.}

\item{...}{additional parameters}
}
\description{
Print information about a GPR model fit, as produced by \code{\link{gaussianProcessRegression}}.
}
\keyword{internal}
