% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{scale01}
\alias{scale01}
\title{Scaling to [0,1]}
\usage{
scale01(x)
}
\arguments{
\item{x}{a vector of values to be scaled}
}
\value{
a vector of scaled values
}
\description{
Scale a set of values to the interval from zero to one.
}
\examples{
scale01(1:10)

}
\keyword{internal}
